requirements.txt
aioredis==1.3.1
alabaster==0.7.8
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.4
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
blackfriday==1.0.3
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==3.3.1
cupshelpers==1.0
cycler==0.10.0
daphne==3.0.1
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.9
django-appconf==1.0.4
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filter==2.4.0
django-googledrive-storage==1.5.2
djangorestframework==3.12.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
feedparser==5.2.1
filelock==3.0.12
future==0.18.2
gitsome==0.8.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
hiredis==1.1.0
httpie==2.2.0
httplib2==0.18.1
hyperlink==21.0.0
idna==2.8
imagesize==1.2.0
incremental==17.5.0
iniconfig==1.0.1
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Jinja2==2.10.1
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.4.0
msgpack==1.0.2
netifaces==0.10.4
numpy==1.20.2
numpydoc==0.7.0
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
packaging==20.4
panda3d==1.10.8
pandas==1.2.3
paramiko==2.6.0
parso==0.7.1
pbr==5.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
prompt-toolkit==3.0.5
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.0.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.6
scipy==1.6.2
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
Sphinx==1.8.5
sqlparse==0.3.1
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
toml==0.10.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.12.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.29
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.12.1
xkit==0.0.0
xonsh==0.9.13
zope.interface==5.2.0
